19: add project name to allow more than one pipeline per host (#21)
* add project name

To keep users separated (multiple pipelines on a single host) need to introduce project names. Using the basename of the envs used to start the pipeline as the project name. It may cause grief later because of how people pick their env names, but it is a good start.

* update env/alsMT

When a user wants to run more than one pipeline on a host, then they must also define EXCALIBUR_USER in their environment configuration.