fix ops logging problem
When workers attach to a pipeline, they get all of the configuration - port numbers included - to match the server. With ops being mapped from 9090 inside the container to 8080 outside the container, the worker was getting the wrong port numbers. Simply moved the internal ports to 8080-8085 so that the workers get the correct port numbers. Since ops is the only pipeline using distributed workers, private pipelines should not be noticably effected.