89: fix reset.sh (#93)
* WIP: unsatisfactory fix

The current fix expands beyond the reset. It is a touch unsatisfactory in that it requires a setting that is difficult with the hijacked variable being used for the host. Need to think of a solution for this because do not like the current one.

* more satisfactory fix

Makes ops another user pipeline from this perspective. It allows for a more consistent usage of envs for all of the pp_* scripts.

* use a patched version of dawgie