134: make collector visible (#142)
* update configuration

Could not access splunk because it was not using https. Added the certificates so that they can be resolved correctly.

Currently every time we have to restart the splunk server, all data is lost. Maybe not so bad because it looks like it will just run forever, but moved the data to the local disk for persistence.