Issue 160: fix up metrics.performance (#161)
* independently track modules

Because metrics.performance derives its list of known modules from dawgie.db.metrics(), it has to keep a persistent stash of them to generate the state vectors correctly for dawgie. It is not user friendly nor editable. It is simply moving code that would have to be update frequently to a pickle file.