Issue 166: update viewing metric information (#169)
* network robustness

As a non sequitur fix, update target.scrape to try a few times when downloading a file with a 2**interation wait window to give the network time to repair itself.

* moved metrics to console

* updated views

Did not like having a state vector per algorithm in the ESP tree. So, used a pull down menu to select it instead. Once plot lets the user walk through the 30 some algorithms known to the system.

Added a data cube of what has run. Bokeh does not do isometric views, so did a top down plane looking at the max run id. Hover over the value in the top view, and the side and front views will change to show the planes from the side and front at that point in the cube. Takes some abstract geometry gymnastics but that is what keeps our brains young.