173 tp profile 2d gettau (#181)
* Fixed verbose mode in Cerberus

* update targetlist (#174)

* more G141 list cleaning

* include Vmag,Zmag in non-mandatory param list
* load in Vmag,Zmag from exoplanet archive

* remove WD 1856+534 from tier-2 Ariel list

* 99 jwst in new code (#176)

* Uncal JWST

* JWST update

* Fix to Cerberus verbose

* Issue 166: update viewing metric information (#169)

* network robustness

As a non sequitur fix, update target.scrape to try a few times when downloading a file with a 2**interation wait window to give the network time to repair itself.

* moved metrics to console

* updated views

Did not like having a state vector per algorithm in the ESP tree. So, used a pull down menu to select it instead. Once plot lets the user walk through the 30 some algorithms known to the system.

Added a data cube of what has run. Bokeh does not do isometric views, so did a top down plane looking at the max run id. Hover over the value in the top view, and the side and front views will change to show the planes from the side and front at that point in the cube. Takes some abstract geometry gymnastics but that is what keeps our brains young.

* 170: target.scape network failures (#171)

* Update core.py

Better network failure messaging

* Adding original errmastq in the log

Adding necessary infos in the logs to better understand the problem
---------

Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>
Co-authored-by: groudier <gael.m.roudier@jpl.nasa.gov>

* update targetlist (#174)

* add new jwst targets to targetlist
* more G141 list cleaning

* include Vmag,Zmag in non-mandatory param list
* load in Vmag,Zmag from exoplanet archive

* remove WD 1856+534 from tier-2 Ariel list

* Scores in JWST

* CIs

* Commenting test lines

* CIs

* Cis

---------

Co-authored-by: al-niessner <1130658+al-niessner@users.noreply.github.com>
Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>
Co-authored-by: Geoffrey Bryden <gbryden@users.noreply.github.com>

* sy_vmag_err1 should be sy_vmagerr1

* Fits file handling (#179)

* Fits file handling

* Dropout

* full 2D atm

* style

* lint

* CI

---------

Co-authored-by: Luke Lamitina <lamitina@mentor3.jpl.nasa.gov>
Co-authored-by: Geoffrey Bryden <gbryden@users.noreply.github.com>
Co-authored-by: al-niessner <1130658+al-niessner@users.noreply.github.com>
Co-authored-by: Al Niessner <Al.Niessner@xxx.xxx>