203: testing fails in CI (#204)
* update the jobs and workflows

First on the agenda was to move to the latest checkout and python-setup.

Second was to remove torch since it is a giant 6 GB package to install it had to be using up some disk space. Turns out that torch was part of the problem. There may still be more, but torch solved it for now.

Despite dawgie ignoring classifier, had to remove torch and all of its references from excalibur.classifier.core. Use this commit to understand how to undo the code removal if classifier ever is enabled.